AAE-25392 Improving service class name
