+++ lib/core/src/lib/auth/guard/auth-guard-ecm.service.ts
@@ -17,11 +17,11 @@
 
 import { ActivatedRouteSnapshot, CanActivateFn, RouterStateSnapshot } from '@angular/router';
 import { AuthenticationService } from '../services/authentication.service';
-import { AuthGuardBaseService } from './auth-guard-base';
+import { AuthGuardService } from './auth-guard.service';
 import { inject } from '@angular/core';
 
 export const AuthGuardEcm: CanActivateFn = async (_: ActivatedRouteSnapshot, state: RouterStateSnapshot): Promise<boolean> => {
-    const authGuardBaseService = inject(AuthGuardBaseService);
+    const authGuardBaseService = inject(AuthGuardService);
     const authenticationService = inject(AuthenticationService);
 
     if (authenticationService.isLoggedIn() && authenticationService.isOauth() && authGuardBaseService.isLoginFragmentPresent()) {
